+++ src/types.d.ts
@@ -48,8 +48,10 @@ declare global {
     offerVariantURLTemplate?: string;
     attributeOverrides?: AttributeOverrides;
     siteOverrides?: Record<string, Record<string, unknown>>;
+    imageParams?: Record<string, string>;
 
     confMap: ConfigMap;
+    confMapStr: string;
   }
 
   export interface Env {
@@ -58,6 +60,7 @@ declare global {
 
     // KV namespaces
     CONFIGS: KVNamespace<string>;
+    KEYS: KVNamespace<string>;
 
     [key: string]: string | KVNamespace<string> | R2Bucket;
   }
@@ -74,6 +77,7 @@ declare global {
     attributes: {
       htmlTemplate?: HTMLTemplate;
       jsonTemplate?: JSONTemplate;
+      key?: string;
       [key: string]: any;
     }
   }